refactor: centralize dominant color logic
